+++ src/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+[Fixed]
+- error handling bug introduced in 0.9.5
+
 ## Version 0.9.5
 (!) WARNING: If you're Humble subscriber, plugin reconnection is needed to sync subscriptions again (!)